Ensure all `_internal` loggers use `SafeLogger`
