+++ src/prefect/_internal/concurrency/__init__.py
@@ -37,31 +37,6 @@ Instead, a simple API is exposed in the  `concurrency.api` module. The API is sp
 respectively.
 """
 
+import prefect._internal.logger
 
-import logging
-
-
-class SafeLogger(logging.Logger):
-    """
-    A logger with extensions for safe emission of logs in our concurrency tooling.
-    """
-
-    def isEnabledFor(self, level: int):
-        # Override `logger.isEnabledFor` to avoid taking a logging lock which can cause
-        # deadlocks during complex concurrency handling
-        from prefect.settings import PREFECT_LOGGING_INTERNAL_LEVEL
-
-        return level >= logging._nameToLevel[PREFECT_LOGGING_INTERNAL_LEVEL.value()]
-
-    def _log(self, *args, **kwargs):
-        from prefect._internal.concurrency.timeouts import shield
-
-        # Prevent interrupts from firing while emitting a log
-        with shield():
-            super()._log(*args, **kwargs)
-
-
-# Use `getLogger` to retain `logger.Manager` behavior
-logger = logging.getLogger("prefect._internal.concurrency")
-# Update the class to inject patched behavior
-logger.__class__ = SafeLogger
+logger = prefect._internal.logger.logger.getChild("concurrency")
